+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed docs theme to 'furo'
+- chore: updated template variables inside `packer.windows`
 
 ## [1.2.0] - 26 Mar 2022